subgroup2/requirements.txt
aiofiles
fastapi
gstreamer-python @ git+https://github.com/jackersson/gstreamer-python.git#egg=gstreamer-python
